third_party/blink/web_tests/http/tests/security/upgrade-insecure-requests/https-header-auxiliary.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
    if (window.testRunner)
      testRunner.setPopupBlockingEnabled(false);

    var t = async_test("Verify that nested documents receive an 'HTTPS' header.");
    window.addEventListener("message", t.step_func(function (message) {
      assert_equals(message.data.header, "1");
      t.done();
    }));

    window.open("resources/post-https-header.pl");
  </script>
</head>
<body>
</body>
</html>
